Safari-611.1.10/LayoutTests/css3/calc/cross-fade-calc.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Test: calc() function in cross-fade</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
    #test {
        background-image: cross-fade(url(dummy://example.png), url(dummy://example.png), calc(35% * 2));
    }
</style>
<div id="test"></div>
<script>
    const div = document.querySelector("#test");
    const expected = 'cross-fade(url("dummy://example.png"), url("dummy://example.png"), 0.7)'
    test(function() {
        assert_equals(getComputedStyle(div).backgroundImage, expected);
    }, "calc() in cross-fade");
</script>
